implementation done, testing left

+++ src/se/kth/IV1350/seminarFour/model/ItemNameQuantityPrice.java
@@ -3,8 +3,8 @@ package se.kth.IV1350.seminarFour.model;
 public class ItemNameQuantityPrice {
     private String name;
     private int quantity;
-    private int price;
-        ItemNameQuantityPrice(String name, int quantity, int price){
+    private double price;
+        ItemNameQuantityPrice(String name, int quantity, double price){
             this.name = name;
             this.quantity = quantity;
             this.price = price;
